statedb: Fix read-after-write bug
[ upstream commit 52944a032150e8040869c20bead7ef05ae32dbe3 ]

[ backporter's note: minor conflicts due to different variables' names]

The optimization in *txn to hold the last used index read transaction
to avoid a hash map lookup was broken as the write txn used for reading
was cloned, leading to not being able to read any of the future writes
in the same transaction.

Benchmark show that this optimization does not actually help, so solve
the problem by removing the optimization.

Before:
BenchmarkDB_RandomLookup-8       1000000              1471 ns/op
After:
BenchmarkDB_RandomLookup-8       1000000              1485 ns/op

Fixes: d0d4d46992bf ("statedb: Store index unique info in the index tree entry")
Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>